experiment/update-mqtt

+++ api/DTOs/Mqtt/RoomSettingsDto.cs
@@ -3,9 +3,9 @@
     public class RoomSettingsDto
     {
         public string Source { get; set; } = "server";
-        public int TargetTemperature { get; set; }
-        public int HumidityThreshold { get; set; }
-        public int HumidityMax { get; set; }
+        public double TargetTemperature { get; set; }
+        public double HumidityThreshold { get; set; }
+        public double HumidityMax { get; set; }
         public int Toggle { get; set; }
         public string Topic { get; set; }
     }
